refactor: unify bytecode serialization to u32 for cross-platform compatibility
- Changed all OpCode indices and offsets from usize to u32.
- Updated Chunk structure to use u32 for line numbers.
- Refactored AST and Parser to use u32 for line tracking.
- Updated Compiler to emit u32-based bytecode.
- Fixed index out of bounds panic in WASM runtime caused by #[serde(skip)] on promoted constants.
- Unified WASM targets on the WASI runner stub to eliminate wasm-bindgen placeholder issues.

+++ crates/matchbox-compiler/src/ast/mod.rs
@@ -1,11 +1,11 @@
 #[derive(Debug, Clone, PartialEq)]
 pub struct Statement {
     pub kind: StatementKind,
-    pub line: usize,
+    pub line: u32,
 }
 
 impl Statement {
-    pub fn new(kind: StatementKind, line: usize) -> Self {
+    pub fn new(kind: StatementKind, line: u32) -> Self {
         Self { kind, line }
     }
 }
@@ -85,11 +85,11 @@ pub enum ClassMember {
 #[derive(Debug, Clone, PartialEq)]
 pub struct Expression {
     pub kind: ExpressionKind,
-    pub line: usize,
+    pub line: u32,
 }
 
 impl Expression {
-    pub fn new(kind: ExpressionKind, line: usize) -> Self {
+    pub fn new(kind: ExpressionKind, line: u32) -> Self {
         Self { kind, line }
     }
 }
