+++ crates/matchbox-vm/src/vm/shape.rs
@@ -1,10 +1,10 @@
 use std::collections::HashMap;
 
-pub type ShapeId = usize;
+pub type ShapeId = u32;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct Shape {
-    pub fields: HashMap<String, usize>,
+    pub fields: HashMap<String, u32>,
     pub transitions: HashMap<String, ShapeId>,
 }
 
@@ -35,27 +35,27 @@ impl ShapeRegistry {
 
     pub fn transition(&mut self, current_id: ShapeId, field_name: &str) -> ShapeId {
         // 1. Check if transition already exists
-        if let Some(&next_id) = self.shapes[current_id].transitions.get(field_name) {
+        if let Some(&next_id) = self.shapes[current_id as usize].transitions.get(field_name) {
             return next_id;
         }
 
         // 2. Create new shape based on current
-        let mut new_shape = self.shapes[current_id].clone();
+        let mut new_shape = self.shapes[current_id as usize].clone();
         new_shape.transitions.clear(); // Transitions are specific to the path taken
         
-        let new_index = new_shape.fields.len();
+        let new_index = new_shape.fields.len() as u32;
         new_shape.fields.insert(field_name.to_string(), new_index);
         
-        let new_id = self.shapes.len();
+        let new_id = self.shapes.len() as u32;
         self.shapes.push(new_shape);
         
         // 3. Record transition in the parent shape
-        self.shapes[current_id].transitions.insert(field_name.to_string(), new_id);
+        self.shapes[current_id as usize].transitions.insert(field_name.to_string(), new_id);
         
         new_id
     }
 
-    pub fn get_index(&self, shape_id: ShapeId, field_name: &str) -> Option<usize> {
-        self.shapes[shape_id].fields.get(field_name).copied()
+    pub fn get_index(&self, shape_id: ShapeId, field_name: &str) -> Option<u32> {
+        self.shapes[shape_id as usize].fields.get(field_name).copied()
     }
 }
